LayoutTests/http/tests/websocket/tests/hixie76/close-before-open-expected.txt

CONSOLE MESSAGE: WebSocket is closed before the connection is established.
Test if Web Socket fires close event when WebSocket is opened and closed fore open event is received.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

closed
PASS successfullyParsed is true

TEST COMPLETE
